moved all componets prepare for trouble and make it double, to protect the world from devestation...
