+++ shift-scan/package.json
@@ -14,15 +14,12 @@
     "fetch": "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/export-db.ts"
   },
   "dependencies": {
-    "@auth/prisma-adapter": "^2.2.0",
     "@prisma/client": "^5.15.0",
     "bcryptjs": "^2.4.3",
-    "js-cookie": "^3.0.5",
     "jsonwebtoken": "^9.0.2",
     "next": "^14.2.3",
     "next-auth": "^4.24.7",
     "next-intl": "^3.14.1",
-    "nookies": "^2.5.2",
     "qr-scanner": "^1.4.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -31,7 +28,6 @@
   "devDependencies": {
     "@types/bcryptjs": "^2.4.6",
     "@types/cookie": "^0.6.0",
-    "@types/js-cookie": "^3.0.6",
     "@types/jsonwebtoken": "^9.0.6",
     "@types/node": "^20.14.2",
     "@types/react": "^18.3.3",
