+++ shift-scan/prisma/seed.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from "@prisma/client";
-import workerData from "../json/worker-data.json";
+import workerData from "@/data/json/worker-data.json";
 import { hash } from "bcryptjs";
 
 const prisma = new PrismaClient();
