+++ shift-scan/src/app/clock/Qr/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React, { useEffect } from 'react';
 import { useTranslations } from 'next-intl';
-import QrReader from '../../../../components/clock/qrReader';
+import QrReader from '@/components/clock/qrReader';
 import { useRouter } from 'next/navigation';
 import { setAuthStep } from '@/app/api/auth';
