+++ shift-scan/src/app/clock/success/page.tsx
@@ -2,10 +2,10 @@
 import React, { useState, useEffect } from 'react';
 import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/navigation';
-import { useScanData } from '../../../../components/context/ScannedJobSIte';
-import { useSavedCostCode } from '../../../../components/context/SavedCostCode';
-import RedirectAfterDelay from '../../../../components/clock/redirectAfterDelay';
-import useBeforeUnload from '../../../../components/app/refreshWarning';
+import { useScanData } from '@/app/context/ScannedJobSIte';
+import { useSavedCostCode } from '@/app/context/SavedCostCode';
+import RedirectAfterDelay from '@/components/clock/redirectAfterDelay';
+import useBeforeUnload from '@/components/app/refreshWarning';
 import { clearAuthStep, getAuthStep, isAuthenticated, setAuthStep } from '@/app/api/auth';
 
 const SuccessPage: React.FC = () => {
