+++ shift-scan/src/components/UI/clock-out.jsx
