+++ shift-scan/src/components/app/banner.tsx
