+++ shift-scan/src/components/app/hoursDisplay.tsx
