+++ shift-scan/src/components/app/refreshWarning.ts
