+++ shift-scan/src/components/authentication/serverRender.tsx
