+++ shift-scan/src/components/authentication/sign-out-btn.tsx
