+++ shift-scan/src/components/clock.tsx
