+++ shift-scan/src/components/clock/redirectAfterDelay.ts
