+++ shift-scan/src/components/clock/searchbar.tsx
