+++ shift-scan/src/components/localeCheckBox.js
